Ensure files are closed and filter out expected warnings
Turns out we were leaving some files open during tests. In addition,
every time we load a wcs during these tests we get a warning that the
number of WCS axes does not match the number of image axes, which is
on purpose -- we want a full-frame WCS but do not want to have to carry
around a full frame image in the repo.